requirements.txt
# Needed
pytest
python-decouple

# Most likely needed
pandas
openpyxl
ipykernel
sqlalchemy
